Change: -all events listner change from rx to android events - delete opportunity of multiple translations - tr layout: delete inner LinearLayout, add progress

+++ app/src/main/java/ru/sinjvf/testtranslate/main/pages/FavoritesPresenter.java
@@ -29,7 +29,9 @@ public class FavoritesPresenter extends HistoryFavoritesPresenter {
         super.clickFavorite(translation, ischecked, position);
         //translations now not favorite remove from list
         if (!isViewAttached())return;
-        getView().deleteItem(position);
+        if(!ischecked) {
+            getView().deleteItem(position);
+        }
     }
